Split remote_desktop_tab.py into a gui/remote_desktop/ subpackage
The single file had ballooned to 1111 lines after the connection-card
redesign — well over CLAUDE.md's 750-line cap. Extract by responsibility:

- _helpers.py        shared utilities (_t, key/button maps, scroll/key
                     event helpers, SSL-context factories, _StatusBadge,
                     _CollapsibleSection)
- frame_display.py   _FrameDisplay widget + drag-drop handling
- host_panel.py      _HostPanel
- viewer_panel.py    _ViewerPanel + _FileSendThread
- tab.py             RemoteDesktopTab outer container
- __init__.py        re-exports the names main_widget / tests use

Largest file is now viewer_panel.py at ~480 lines. The old
remote_desktop_tab.py becomes a thin shim that re-exports the same
public names so existing import paths (used by main_widget and the
GUI integration tests) keep working.
